--- a/tt_eager/tt_dnn/op_library/nlp_tms/nlp_tms.hpp
+++ b/tt_eager/tt_dnn/op_library/nlp_tms/nlp_tms.hpp
@@ -116,21 +116,28 @@ inline std::vector<Tensor> nlp_create_qkv_heads(
     const bool transpose_k_heads,
     const MemoryConfig& mem_config
 ) {
-    const uint32_t num_kv_heads_val = num_kv_heads.value_or(num_heads);
-
-    // Infer head_dim
-    uint32_t head_dim;
-    if (input_tensor_kv.has_value()) {
-        TT_FATAL(input_tensor.get_legacy_shape()[3] % num_heads == 0, "Unsupported input shape");
-        TT_FATAL(input_tensor_kv.value().get_legacy_shape()[3] % (2 * num_kv_heads_val) == 0, "Unsupported input shape");
-        head_dim = input_tensor.get_legacy_shape()[3] / num_heads;
-        TT_FATAL(input_tensor_kv.value().get_legacy_shape()[3] / (2 * num_kv_heads_val) == head_dim, "Head dims must be the same for Q and K, V");
-    } else {
-        TT_FATAL(input_tensor.get_legacy_shape()[3] % (num_heads + 2 * num_kv_heads_val) == 0, "Unsupported input shape");
-        head_dim = input_tensor.get_legacy_shape()[3] / (num_heads + 2 * num_kv_heads_val);
-    }
-
-    return operation::run(NlpCreateHeads{num_heads, num_kv_heads_val, head_dim, transpose_k_heads, mem_config}, {input_tensor}, {input_tensor_kv});
+    std::vector<Tensor> output_tensors = {Tensor(operation::get_workers_for_op_output({input_tensor}, {input_tensor_kv})),
+                                          Tensor(operation::get_workers_for_op_output({input_tensor}, {input_tensor_kv})),
+                                          Tensor(operation::get_workers_for_op_output({input_tensor}, {input_tensor_kv}))};
+    operation::launch_op(
+        [num_heads, num_kv_heads, transpose_k_heads, mem_config] (std::vector<Tensor> input_tensors, const std::vector<std::optional<const Tensor>>& optional_input_tensors) mutable -> std::vector<Tensor> {
+            auto& input_tensor = input_tensors.at(0);
+            auto& input_tensor_kv = optional_input_tensors.at(0);
+            const uint32_t num_kv_heads_val = num_kv_heads.value_or(num_heads);
+            // Infer head_dim
+            uint32_t head_dim;
+            if (input_tensor_kv.has_value()) {
+                TT_FATAL(input_tensor.get_legacy_shape()[3] % num_heads == 0, "Unsupported input shape");
+                TT_FATAL(input_tensor_kv.value().get_legacy_shape()[3] % (2 * num_kv_heads_val) == 0, "Unsupported input shape");
+                head_dim = input_tensor.get_legacy_shape()[3] / num_heads;
+                TT_FATAL(input_tensor_kv.value().get_legacy_shape()[3] / (2 * num_kv_heads_val) == head_dim, "Head dims must be the same for Q and K, V");
+            } else {
+                TT_FATAL(input_tensor.get_legacy_shape()[3] % (num_heads + 2 * num_kv_heads_val) == 0, "Unsupported input shape");
+                head_dim = input_tensor.get_legacy_shape()[3] / (num_heads + 2 * num_kv_heads_val);
+            }
+            return operation::run(NlpCreateHeads{num_heads, num_kv_heads_val, head_dim, transpose_k_heads, mem_config}, {input_tensor}, {input_tensor_kv});
+        }, {input_tensor}, output_tensors, {input_tensor_kv});
+    return output_tensors;
 }
 inline Tensor nlp_concat_heads(const Tensor &input_tensor_a, const MemoryConfig& mem_config) {
     std::vector<Tensor> output_tensors = {Tensor(operation::get_workers_for_op_output({input_tensor_a}))};