@cglagovichTT 2024-05-02
With async on:
```
2024-05-02 22:09:05.744 | INFO     | models.demos.t3000.llama2_70b.demo.demo:latency_printout:194 - Overall throughput: 28.1 ms @ 35.6 tokens/s
2024-05-02 22:09:05.744 | INFO     | models.demos.t3000.llama2_70b.demo.demo:latency_printout:197 - Tokens per second per user: 1.1 tokens/s/u
2024-05-02 22:09:05.744 | INFO     | models.demos.t3000.llama2_70b.demo.demo:latency_printout:198 - User latency: 898.7 ms @ 1.1 tokens/s
```

Without async:
```
2024-05-02 22:31:34.335 | INFO     | models.demos.t3000.llama2_70b.demo.demo:latency_printout:194 - Overall throughput: 46.2 ms @ 21.7 tokens/s
2024-05-02 22:31:34.335 | INFO     | models.demos.t3000.llama2_70b.demo.demo:latency_printout:197 - Tokens per second per user: 0.7 tokens/s/u
2024-05-02 22:31:34.335 | INFO     | models.demos.t3000.llama2_70b.demo.demo:latency_printout:198 - User latency: 1477.6 ms @ 0.7 tokens/s
```