@cglagovichTT 2024-05-02
Release mode, sync:
```
2024-05-02 22:53:36.708 | INFO     | models.demos.t3000.llama2_70b.demo.demo:latency_printout:194 - Overall throughput: 25.1 ms @ 39.8 tokens/s
2024-05-02 22:53:36.708 | INFO     | models.demos.t3000.llama2_70b.demo.demo:latency_printout:197 - Tokens per second per user: 1.2 tokens/s/u
2024-05-02 22:53:36.708 | INFO     | models.demos.t3000.llama2_70b.demo.demo:latency_printout:198 - User latency: 804.3 ms @ 1.2 tokens/s
```

async
```
2024-05-02 23:04:25.067 | INFO     | models.demos.t3000.llama2_70b.demo.demo:latency_printout:194 - Overall throughput: 13.2 ms @ 76.0 tokens/s
2024-05-02 23:04:25.067 | INFO     | models.demos.t3000.llama2_70b.demo.demo:latency_printout:197 - Tokens per second per user: 2.4 tokens/s/u
2024-05-02 23:04:25.067 | INFO     | models.demos.t3000.llama2_70b.demo.demo:latency_printout:198 - User latency: 421.2 ms @ 2.4 tokens/s
```